+++ http4k-format/core/src/test/kotlin/org/http4k/format/AutoMarshallingJsonContract.kt
@@ -59,4 +59,5 @@ abstract class AutoMarshallingJsonContract(marshaller: AutoMarshalling) : AutoMa
     fun `handles unit`() {
         customMarshaller().asA<Unit>("{}")
     }
+
 }