+++ http4k-format/moshi/src/main/kotlin/org/http4k/format/adapters.kt
@@ -3,7 +3,9 @@ package org.http4k.format
 import com.squareup.moshi.JsonAdapter
 import com.squareup.moshi.Moshi
 import com.squareup.moshi.Types
+import org.http4k.events.Event
 import java.lang.reflect.Type
+import kotlin.reflect.KClass
 
 /**
  * Convenience class to create Moshi Adapter Factory
@@ -31,3 +33,33 @@ inline fun <reified K> JsonAdapter<K>.asFactory() = SimpleMoshiAdapterFactory(K:
  */
 inline fun <reified T : JsonAdapter<K>, reified K> Moshi.Builder.addTyped(fn: T): Moshi.Builder =
     add(K::class.java, fn)
+
+
+/**
+ * This adapter factory will capture ALL instances of a particular superclass/interface.
+ */
+abstract class IsAnInstanceOfAdapter<T : Any>(private val clazz: KClass<T>): JsonAdapter.Factory {
+    override fun create(type: Type, annotations: Set<Annotation>, moshi: Moshi) =
+        with(Types.getRawType(type)) {
+            when {
+                isA(clazz.java) -> moshi.adapter(clazz.java)
+                else -> null
+            }
+        }
+
+    private fun Class<*>?.isA(testCase: Class<*>): Boolean =
+        this?.let { testCase != this && testCase.isAssignableFrom(this) } ?: false
+}
+
+/**
+ * These adapters are the edge case adapters for dealing with Moshi
+ */
+
+object ThrowableAdapter : IsAnInstanceOfAdapter<Throwable>(Throwable::class)
+
+object CollectionEdgeCasesAdapter : IsAnInstanceOfAdapter<Map<*, *>>(Map::class)
+
+object EventAdapter : JsonAdapter.Factory {
+    override fun create(p0: Type, p1: MutableSet<out Annotation>, p2: Moshi) =
+        if (p0.typeName == Event::class.java.typeName) p2.adapter(Any::class.java) else null
+}